refactor: abstract out webpack config logic
